Release v2.8

+++ scripts/document.sh
@@ -4,31 +4,26 @@ set -e
 set -o pipefail
 set -u
 
-if [ -z `which jazzy` ]; then
-    echo "Installing jazzy…"
-    gem install jazzy
-    if [ -z `which jazzy` ]; then
-        echo "Unable to install jazzy. See https://github.com/mapbox/mapbox-gl-native-ios/blob/master/platform/ios/INSTALL.md"
-        exit 1
-    fi
+bundle check || bundle install
+
+SOURCEKITTEN_PATH="$(dirname "$(bundle exec gem which jazzy)")/../bin/sourcekitten"
+
+if [[ -z $(which "${SOURCEKITTEN_PATH}") ]]; then
+    echo "Unable to locate SourceKitten. See installation instructions at https://github.com/jpsim/SourceKitten#installation"
+    exit 1
 fi
-
-
-OUTPUT=${OUTPUT:-documentation}
 
 BRANCH=$( git describe --tags --match=v*.*.* --abbrev=0 )
 SHORT_VERSION=$( echo ${BRANCH} | sed 's/^v//' )
 RELEASE_VERSION=$( echo ${SHORT_VERSION} | sed -e 's/-.*//' )
 MINOR_VERSION=$( echo ${SHORT_VERSION} | grep -Eo '^\d+\.\d+' )
 
+OUTPUT=${OUTPUT:-${SHORT_VERSION:-documentation}}
+
 DEFAULT_THEME="docs/theme"
 THEME=${JAZZY_THEME:-$DEFAULT_THEME}
 
 BASE_URL="https://docs.mapbox.com/ios"
-
-# Link to directions documentation
-DIRECTIONS_VERSION=$(grep 'mapbox-directions-swift' Cartfile.resolved | grep -oE '"v.+?"' | grep -oE '[^"v]+')
-DIRECTIONS_SYMBOLS="AttributeOptions|CoordinateBounds|Directions|DirectionsCredentials|DirectionsOptions|DirectionsPriority|DirectionsProfileIdentifier|DirectionsResult|Intersection|Lane|LaneIndication|MapMatchingResponse|Match|MatchOptions|RoadClasses|Route|RouteLeg|RouteOptions|RouteResponse|RouteStep|SpokenInstruction|Tracepoint|VisualInstruction|VisualInstruction.Component|VisualInstruction.Component.ImageRepresentation|VisualInstruction.Component.TextRepresentation|VisualInstructionBanner|Waypoint"
 
 rm -rf ${OUTPUT}
 mkdir -p ${OUTPUT}
@@ -40,42 +35,40 @@ mkdir -p /tmp/mbnavigation/
 README=/tmp/mbnavigation/README.md
 cp docs/cover.md "${README}"
 perl -pi -e "s/\\$\\{MINOR_VERSION\\}/${MINOR_VERSION}/" "${README}"
+perl -pi -e "s/\\$\\{SHORT_VERSION\\}/${SHORT_VERSION}/" "${README}"
 # http://stackoverflow.com/a/4858011/4585461
 echo "## Changes in version ${RELEASE_VERSION}" >> "${README}"
 sed -n -e '/^## /{' -e ':a' -e 'n' -e '/^## /q' -e 'p' -e 'ba' -e '}' CHANGELOG.md >> "${README}"
+    
+PROJECT="MapboxNavigation-SPM.xcodeproj"
+DESTINATION="generic/platform=iOS"
+"${SOURCEKITTEN_PATH}" doc --module-name MapboxCoreNavigation -- -project "${PROJECT}" -destination "${DESTINATION}" -scheme MapboxCoreNavigation > core.json
+"${SOURCEKITTEN_PATH}" doc --module-name MapboxNavigation -- -project "${PROJECT}" -destination "${DESTINATION}" -scheme MapboxNavigation > ui.json
 
-# Blow away any includes of MapboxCoreNavigation, because
-# MapboxNavigation-Documentation.podspec gloms the two targets into one.
-# https://github.com/mapbox/mapbox-navigation-ios/issues/2363
-find Sources/Mapbox{Core,}Navigation/ -name '*.swift' -exec \
-    perl -pi -e 's/\bMapboxCoreNavigation\b/MapboxNavigation/' {} \;
-find Sources/Mapbox{Core,}Navigation/ -name '*.[hm]' -exec \
-    perl -pi -e 's/([<"])MapboxCoreNavigation\b/$1MapboxNavigation/' {} \;
-
-# Blow away any platform-based availability attributes, since everything is
-# compatible enough to be documented.
-# https://github.com/mapbox/mapbox-navigation-ios/issues/1682
-find Sources/Mapbox{Core,}Navigation/ -name '*.swift' -exec \
-    perl -pi -e 's/\@available\s*\(\s*iOS \d+.\d,.*?\)//' {} \;
-
-jazzy \
-    --podspec MapboxNavigation-Documentation.podspec \
+bundle exec jazzy \
     --config docs/jazzy.yml \
     --sdk iphonesimulator \
-    --module-version ${SHORT_VERSION} \
     --github-file-prefix "https://github.com/mapbox/mapbox-navigation-ios/tree/${BRANCH}" \
     --readme ${README} \
     --documentation="docs/guides/*.md" \
     --root-url "${BASE_URL}/navigation/api/${RELEASE_VERSION}/" \
     --theme ${THEME} \
     --output ${OUTPUT} \
-    --module_version ${RELEASE_VERSION}
+    --module_version ${RELEASE_VERSION} \
+    --sourcekitten-sourcefile core.json,ui.json \
+    2>&1 | tee docs.output
+    
+rm core.json ui.json
+    
+if egrep -e "(WARNING)|(USR)" docs.output; then
+    echo "Please eliminate Jazzy warnings"
+    exit 1
+fi
+    
+rm docs.output
 
-REPLACE_REGEXP='s/MapboxNavigation\s+(Docs|Reference)/Mapbox Navigation SDK for iOS $1/, '
-REPLACE_REGEXP+="s/<span class=\"kt\">(${DIRECTIONS_SYMBOLS})<\/span>/<span class=\"kt\"><a href=\"${BASE_URL//\//\\/}\/directions\/api\/${DIRECTIONS_VERSION}\/Classes\/\$1.html\">\$1<\/a><\/span>/, "
-
-find ${OUTPUT} -name *.html -exec \
-    perl -pi -e "$REPLACE_REGEXP" {} \;
-
+# Link to directions documentation
+DIRECTIONS_VERSION=$(python3 -c "import json; print(list(filter(lambda x:x['package']=='MapboxDirections', json.loads(open('Package.resolved').read())['object']['pins']))[0]['state']['version'])")
+python3 scripts/postprocess-docs.py -b "${BASE_URL}/directions/api/${DIRECTIONS_VERSION}" -d "${OUTPUT}"
 
 echo $SHORT_VERSION > $OUTPUT/latest_version
